Code cleanup. 1. Create DataService. 2. Put EtlController's methods implementation into services. 3. Pack up classes.
