+++ src/main/java/com/tmw/etl/etlapp/exc/NoDataException.java
@@ -1,4 +1,4 @@
-package com.tmw.etl.etlapp;
+package com.tmw.etl.etlapp.exc;
 
 public class NoDataException extends Exception {
     public NoDataException(String msg){
