+++ src/main/resources/application.properties
@@ -1,5 +1,4 @@
 spring.jpa.hibernate.ddl-auto=create
 spring.datasource.url=jdbc:mysql://localhost:3306/etlapp?useSSL=false&useUnicode=yes&characterEncoding=UTF-8&characterSetResults=UTF-8
 spring.datasource.username=root
-spring.datasource.password=wieliczka11
-
+spring.datasource.password=Test123.
